Address TSE issue#2 & refactor the code  - Change .eslintrc.json file to comform microverse standards  - Change named exports to default exprot in the components folder  - Add pages folder and put restaurant pages inside the folder
