+++ src/components/header.js
@@ -1,6 +1,6 @@
-const { getButton } = require('./button');
+const getButton = require('./button');
 
-export const getHeader = () => {
+const getHeader = () => {
   const UIHeader = document.createElement('header');
   UIHeader.id = 'showcase';
   const UIContainerDiv = document.createElement('div');
@@ -23,7 +23,7 @@ laudantium. Quam sapiente a dolorum magnam necessitatibus quis
 tempore facere totam. Dolor, sequi distinctio!`;
   UIShowcaseContentDiv.appendChild(UIShowcasePara);
 
-  const UIShowcaseOrderBtn = getButton('menu', ['btn-order', 'btn', 'btn-primary'], 'Order');
+  const UIShowcaseOrderBtn = getButton.default('menu', ['btn-order', 'btn', 'btn-primary'], 'Order');
   UIShowcaseContentDiv.appendChild(UIShowcaseOrderBtn);
 
   UIShowcaseContainerDiv.appendChild(UIShowcaseContentDiv);
@@ -32,3 +32,5 @@ tempore facere totam. Dolor, sequi distinctio!`;
 
   return UIHeader;
 };
+
+export default getHeader;
